fix for non overlay sessions! d3d12 session chain impl missing, will fail to compile
