+++ crates/bevy_openxr/src/openxr/mod.rs
@@ -30,7 +30,7 @@ pub mod render;
 pub mod resources;
 pub mod session;
 pub mod types;
-pub mod session_create_info_builder;
+pub mod session_create_info_chain;
 
 pub fn add_xr_plugins<G: PluginGroup>(plugins: G) -> PluginGroupBuilder {
     plugins
